lambda/requirements.txt
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.9
boto3==1.35.36
